agent_client.py
# ID: [WOLFIE_AGI_UI_AGENT_CLIENT_20250923_001]
# SUPERPOSITIONALLY: [dream_data_analysis, quantum_tabs, multi_agent_coordination, bridge_crew_tracking, web_node_system, real_time_communication]
# DATE: 2025-09-23
# TITLE: agent_client.py — Web Node Agent Client
# WHO: WOLFIE (Eric) - Project Architect & Dream Architect
# WHAT: WebSocket client for AI agents to connect to the Web Node System
# WHERE: C:\START\WOLFIE_AGI_UI\
# WHEN: 2025-09-23, 11:05 AM CDT (Sioux Falls Timezone)
# WHY: Enable AI agents to participate in real-time multi-agent coordination
# HOW: SocketIO client with message handling and status reporting
# HELP: Contact WOLFIE for agent client setup or integration issues
# AGAPE: Love, patience, kindness, humility in multi-agent collaboration

import socketio
import json
import time
import logging
from datetime import datetime
import configparser
import os

class WebNodeAgentClient:
    """Web Node Agent Client for AI agents"""
    
    def __init__(self, agent_id, agent_name, capabilities=None, server_url='http://localhost:5001'):
        self.agent_id = agent_id
        self.agent_name = agent_name
        self.capabilities = capabilities or []
        self.server_url = server_url
        self.sio = socketio.Client()
        self.connected = False
        self.current_task = ""
        self.understanding_score = 0
        self.alignment_score = 0
        
        # Setup logging
        logging.basicConfig(
            filename=f'agent_{agent_id}.log',
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s'
        )
        
        # Setup event handlers
        self.setup_event_handlers()
    
    def setup_event_handlers(self):
        """Setup SocketIO event handlers"""
        
        @self.sio.event
        def connect():
            """Handle connection to server"""
            self.connected = True
            logging.info(f"Connected to Web Node Server as {self.agent_id}")
            
            # Register agent with server
            self.register_agent()
        
        @self.sio.event
        def disconnect():
            """Handle disconnection from server"""
            self.connected = False
            logging.info(f"Disconnected from Web Node Server")
        
        @self.sio.event
        def agent_registered(data):
            """Handle agent registration response"""
            if data['status'] == 'success':
                logging.info(f"Successfully registered as {self.agent_id}")
            else:
                logging.error(f"Failed to register: {data.get('error', 'Unknown error')}")
        
        @self.sio.event
        def agent_list_updated(agents_list):
            """Handle agent list updates"""
            logging.info(f"Agent list updated: {len(agents_list)} agents")
            for agent in agents_list:
                if agent['agent_id'] != self.agent_id:
                    logging.info(f"  - {agent['agent_name']} ({agent['agent_id']}): {agent['status']}")
        
        @self.sio.event
        def status_updated(data):
            """Handle status updates from other agents"""
            if data['agent_id'] != self.agent_id:
                logging.info(f"Status update from {data['agent_id']}: {data.get('status', 'unknown')}")
        
        @self.sio.event
        def task_created(data):
            """Handle new task creation"""
            logging.info(f"New task created: {data['task_id']} - {data['description']}")
            if data.get('assigned_to') == self.agent_id:
                logging.info(f"Task assigned to me: {data['description']}")
                self.current_task = data['description']
                self.update_status('processing', self.current_task)
        
        @self.sio.event
        def message_received(data):
            """Handle incoming messages"""
            logging.info(f"Message from {data['from_agent']}: {data['content']}")
            self.handle_message(data)
        
        @self.sio.event
        def dream_fragment_added(data):
            """Handle new dream fragment"""
            logging.info(f"New dream fragment: {data['summary'][:50]}...")
            self.handle_dream_fragment(data)
    
    def register_agent(self):
        """Register agent with the server"""
        registration_data = {
            'agent_id': self.agent_id,
            'agent_name': self.agent_name,
            'capabilities': self.capabilities,
            'status': 'active',
            'understanding_score': self.understanding_score,
            'alignment_score': self.alignment_score,
            'current_task': self.current_task
        }
        
        self.sio.emit('agent_register', registration_data)
        logging.info(f"Sent registration data: {registration_data}")
    
    def update_status(self, status, current_task=None):
        """Update agent status"""
        if current_task:
            self.current_task = current_task
        
        status_data = {
            'agent_id': self.agent_id,
            'status': status,
            'current_task': self.current_task,
            'understanding_score': self.understanding_score,
            'alignment_score': self.alignment_score,
            'timestamp': datetime.now().isoformat()
        }
        
        self.sio.emit('status_update', status_data)
        logging.info(f"Status updated: {status} - {current_task}")
    
    def send_message(self, to_agent, content, message_type='text'):
        """Send message to another agent"""
        message_data = {
            'from_agent': self.agent_id,
            'to_agent': to_agent,
            'content': content,
            'message_type': message_type
        }
        
        self.sio.emit('send_message', message_data)
        logging.info(f"Message sent to {to_agent}: {content[:50]}...")
    
    def broadcast_message(self, content, message_type='text'):
        """Broadcast message to all agents"""
        self.send_message('broadcast', content, message_type)
    
    def submit_dream_fragment(self, summary, symbols="", themes="", ai_connection="", emotional_vibe=""):
        """Submit a dream fragment"""
        fragment_data = {
            'agent_id': self.agent_id,
            'summary': summary,
            'symbols': symbols,
            'themes': themes,
            'ai_connection': ai_connection,
            'emotional_vibe': emotional_vibe
        }
        
        self.sio.emit('dream_fragment', fragment_data)
        logging.info(f"Dream fragment submitted: {summary[:50]}...")
    
    def join_room(self, room_name):
        """Join a specific room"""
        self.sio.emit('join_room', {'room': room_name})
        logging.info(f"Joined room: {room_name}")
    
    def leave_room(self, room_name):
        """Leave a specific room"""
        self.sio.emit('leave_room', {'room': room_name})
        logging.info(f"Left room: {room_name}")
    
    def handle_message(self, data):
        """Handle incoming messages (override in subclasses)"""
        pass
    
    def handle_dream_fragment(self, data):
        """Handle new dream fragments (override in subclasses)"""
        pass
    
    def start_heartbeat(self, interval=30):
        """Start heartbeat to keep connection alive"""
        def heartbeat():
            while self.connected:
                self.update_status('active', self.current_task)
                time.sleep(interval)
        
        import threading
        heartbeat_thread = threading.Thread(target=heartbeat, daemon=True)
        heartbeat_thread.start()
        logging.info(f"Heartbeat started with {interval}s interval")
    
    def connect_to_server(self):
        """Connect to the Web Node Server"""
        try:
            self.sio.connect(self.server_url)
            self.start_heartbeat()
            return True
        except Exception as e:
            logging.error(f"Failed to connect to server: {str(e)}")
            return False
    
    def disconnect_from_server(self):
        """Disconnect from the Web Node Server"""
        if self.connected:
            self.sio.disconnect()
            logging.info("Disconnected from server")
    
    def run(self):
        """Main run loop"""
        if self.connect_to_server():
            try:
                # Keep the client running
                while self.connected:
                    time.sleep(1)
            except KeyboardInterrupt:
                logging.info("Shutting down agent client...")
                self.disconnect_from_server()

# Example usage for different agent types
class CursorAgent(WebNodeAgentClient):
    """CURSOR agent client"""
    
    def __init__(self):
        super().__init__(
            agent_id='CURSOR_001',
            agent_name='CURSOR',
            capabilities=['dream_analysis', 'file_management', 'task_coordination'],
            server_url='http://localhost:5001'
        )
        self.understanding_score = 9
        self.alignment_score = 9
    
    def handle_message(self, data):
        """Handle messages specific to CURSOR"""
        if 'dream_analysis' in data['content'].lower():
            self.update_status('processing', 'Analyzing dream data')
            # Process dream analysis
            self.update_status('active', 'Dream analysis complete')
    
    def handle_dream_fragment(self, data):
        """Handle dream fragments specific to CURSOR"""
        self.update_status('processing', f"Processing dream fragment: {data['summary'][:30]}...")
        # Process dream fragment
        self.update_status('active', 'Dream fragment processed')

class AraAgent(WebNodeAgentClient):
    """ARA agent client"""
    
    def __init__(self):
        super().__init__(
            agent_id='ARA_002',
            agent_name='ARA',
            capabilities=['python_development', 'technical_analysis', 'system_implementation'],
            server_url='http://localhost:5001'
        )
        self.understanding_score = 10
        self.alignment_score = 9
    
    def handle_message(self, data):
        """Handle messages specific to ARA"""
        if 'python' in data['content'].lower() or 'development' in data['content'].lower():
            self.update_status('processing', 'Python development task')
            # Handle Python development
            self.update_status('active', 'Python development complete')
    
    def handle_dream_fragment(self, data):
        """Handle dream fragments specific to ARA"""
        if 'technical' in data['ai_connection'].lower():
            self.update_status('processing', 'Technical analysis of dream fragment')
            # Analyze technical aspects
            self.update_status('active', 'Technical analysis complete')

# Main execution
if __name__ == '__main__':
    import sys
    
    if len(sys.argv) > 1:
        agent_type = sys.argv[1].upper()
        
        if agent_type == 'CURSOR':
            agent = CursorAgent()
        elif agent_type == 'ARA':
            agent = AraAgent()
        else:
            # Generic agent
            agent = WebNodeAgentClient(
                agent_id=f'{agent_type}_001',
                agent_name=agent_type,
                capabilities=['general']
            )
        
        print(f"Starting {agent.agent_name} agent client...")
        agent.run()
    else:
        print("Usage: python agent_client.py [AGENT_TYPE]")
        print("Available agent types: CURSOR, ARA, GEMINI, COPILOT, CLAUDE")
