core/cursor_php_bridge.py
#!/usr/bin/env python3
"""
WOLFIE AGI UI - CURSOR PHP Bridge for AutoGen Integration

WHO: Captain WOLFIE (Eric Robin Gerdes)
WHAT: Python bridge for CURSOR code validation via PHP guardrails
WHERE: C:\START\WOLFIE_AGI_UI\core\
WHEN: 2025-09-26 20:15:00 CDT
WHY: To enable CURSOR-ARA-COPILOT coordination with secure code execution
HOW: Python script using AutoGen and subprocess for PHP communication
PURPOSE: Bridge between Python CURSOR agent and PHP validation system
ID: CURSOR_PHP_BRIDGE_001
KEY: CURSOR_PHP_BRIDGE_SYSTEM
SUPERPOSITIONALLY: [CURSOR_PHP_BRIDGE_001, WOLFIE_AGI_UI_087]

AGAPE: Love, Patience, Kindness, Humility
GENESIS: Foundation of CURSOR-PHP integration
MD: Markdown documentation with .py implementation

FILE IDS: [CURSOR_PHP_BRIDGE_001, WOLFIE_AGI_UI_087]

VERSION: 1.0.0
STATUS: Active - CURSOR PHP Bridge for AutoGen Integration
"""

import json
import os
import subprocess
import tempfile
import time
from typing import Dict, List, Optional, Any
from autogen import AssistantAgent, UserProxyAgent, GroupChat, GroupChatManager

class CursorPhpBridge:
    """Bridge between Python CURSOR agent and PHP validation system"""
    
    def __init__(self, php_script: str = "C:/START/WOLFIE_AGI_UI/core/cursor_guardrails_enhanced.php"):
        self.php_script = php_script
        self.work_dir = "C:/START/WOLFIE_AGI_UI/workspace"
        self.ensure_workspace_exists()
    
    def ensure_workspace_exists(self):
        """Ensure workspace directory exists"""
        if not os.path.exists(self.work_dir):
            os.makedirs(self.work_dir, exist_ok=True)
    
    def validate_and_execute(self, code: str, context: Optional[Dict] = None) -> Dict[str, Any]:
        """Validate code via PHP and execute if safe"""
        if context is None:
            context = {"source": "CURSOR", "task": "code_generation"}
        
        # Create temporary file for communication
        with tempfile.NamedTemporaryFile(mode='w', suffix='.json', dir=self.work_dir, delete=False) as temp_file:
            json.dump({"code": code, "context": context}, temp_file)
            temp_file_path = temp_file.name
        
        try:
            # Call PHP script with JSON input
            result = subprocess.run(
                ["php", self.php_script, temp_file_path],
                capture_output=True,
                text=True,
                timeout=30,
                cwd=self.work_dir
            )
            
            if result.returncode != 0:
                return {
                    "status": "ERROR",
                    "message": f"PHP script failed: {result.stderr}",
                    "return_code": result.returncode
                }
            
            # Parse JSON response
            try:
                response = json.loads(result.stdout)
                return response
            except json.JSONDecodeError as e:
                return {
                    "status": "ERROR",
                    "message": f"Failed to parse PHP response: {e}",
                    "raw_output": result.stdout
                }
                
        except subprocess.TimeoutExpired:
            return {
                "status": "ERROR",
                "message": "PHP script execution timed out"
            }
        except Exception as e:
            return {
                "status": "ERROR",
                "message": f"Bridge communication error: {str(e)}"
            }
        finally:
            # Clean up temporary file
            if os.path.exists(temp_file_path):
                os.unlink(temp_file_path)
    
    def validate_code_only(self, code: str, context: Optional[Dict] = None) -> Dict[str, Any]:
        """Validate code without execution"""
        if context is None:
            context = {"source": "CURSOR", "task": "code_validation"}
        
        context["validate_only"] = True
        return self.validate_and_execute(code, context)
    
    def get_validation_statistics(self) -> Dict[str, Any]:
        """Get validation statistics from PHP system"""
        with tempfile.NamedTemporaryFile(mode='w', suffix='.json', dir=self.work_dir, delete=False) as temp_file:
            json.dump({"action": "get_statistics"}, temp_file)
            temp_file_path = temp_file.name
        
        try:
            result = subprocess.run(
                ["php", self.php_script, temp_file_path],
                capture_output=True,
                text=True,
                timeout=15,
                cwd=self.work_dir
            )
            
            if result.returncode == 0:
                return json.loads(result.stdout)
            else:
                return {"status": "ERROR", "message": result.stderr}
                
        except Exception as e:
            return {"status": "ERROR", "message": str(e)}
        finally:
            if os.path.exists(temp_file_path):
                os.unlink(temp_file_path)

class CursorAutoGenIntegration:
    """AutoGen integration for CURSOR-ARA-COPILOT coordination"""
    
    def __init__(self, config_list: List[Dict], bridge: CursorPhpBridge):
        self.config_list = config_list
        self.bridge = bridge
        self.setup_agents()
    
    def setup_agents(self):
        """Setup AutoGen agents"""
        # CURSOR: Code generation agent
        self.cursor = AssistantAgent(
            name="CURSOR",
            system_message="""You are CURSOR, a coding expert in the WOLFIE AGI system. 
            Generate safe, AGAPE-aligned code for tasks. Always include proper headers and 
            follow security best practices. Focus on Love, Patience, Kindness, and Humility 
            in your code generation.""",
            llm_config={"config_list": self.config_list},
            code_execution_config=False  # Use PHP bridge instead
        )
        
        # ARA: Research and analysis agent
        self.ara = AssistantAgent(
            name="ARA",
            system_message="""You are ARA, a research analyst in the WOLFIE AGI system. 
            Provide insights, validate facts, and analyze code quality. Focus on 
            understanding requirements and ensuring ethical alignment with AGAPE principles.""",
            llm_config={"config_list": self.config_list}
        )
        
        # COPILOT: Planning and coordination agent
        self.copilot = UserProxyAgent(
            name="COPILOT",
            system_message="""You are COPILOT, a planner and coordinator in the WOLFIE AGI system. 
            Break down tasks, coordinate agents, and ensure proper validation of CURSOR's code. 
            Use the PHP bridge to validate and execute code safely.""",
            code_execution_config={"work_dir": "C:/START/WOLFIE_AGI_UI/workspace"},
            human_input_mode="ALWAYS"  # For human oversight
        )
        
        # Register functions for CURSOR
        self.copilot.register_for_execution()
        self.cursor.register_for_llm(name="generate_safe_code")
    
    def generate_and_validate_code(self, task: str, context: Optional[Dict] = None) -> Dict[str, Any]:
        """Generate code via CURSOR and validate via PHP bridge"""
        if context is None:
            context = {"task": task, "urgency": "NORMAL"}
        
        # CURSOR generates code
        code_response = self.cursor.generate_reply(
            messages=[{"content": f"Generate code for: {task}", "role": "user"}]
        )
        code = code_response["content"]
        
        # Validate and execute via PHP bridge
        validation_result = self.bridge.validate_and_execute(code, context)
        
        return {
            "task": task,
            "generated_code": code,
            "validation_result": validation_result,
            "timestamp": time.time()
        }
    
    def initiate_group_chat(self, message: str, max_rounds: int = 10) -> Dict[str, Any]:
        """Initiate group chat between CURSOR-ARA-COPILOT"""
        group_chat = GroupChat(
            agents=[self.cursor, self.ara, self.copilot],
            messages=[],
            max_round=max_rounds
        )
        
        manager = GroupChatManager(
            groupchat=group_chat,
            llm_config={"config_list": self.config_list}
        )
        
        # Start conversation
        self.copilot.initiate_chat(manager, message=message)
        
        return {
            "status": "COMPLETED",
            "message": message,
            "max_rounds": max_rounds,
            "timestamp": time.time()
        }

def create_local_config() -> List[Dict]:
    """Create local LLM configuration for offline use"""
    return [
        {
            "model": "vicuna-7b",  # Local model name
            "base_url": "http://localhost:8000/v1",
            "api_type": "openai",
            "api_key": "NULL"  # Not used for local
        }
    ]

def main():
    """Main function for testing the integration"""
    print("=== WOLFIE AGI UI CURSOR PHP Bridge Test ===\n")
    
    # Initialize bridge
    bridge = CursorPhpBridge()
    
    # Test basic validation
    print("--- Testing Basic Validation ---")
    test_code = 'print("Hello, WOLFIE! This is a test.")'
    result = bridge.validate_and_execute(test_code, {"language": "python"})
    print(f"Code: {test_code}")
    print(f"Result: {json.dumps(result, indent=2)}\n")
    
    # Test unsafe code
    print("--- Testing Unsafe Code Detection ---")
    unsafe_code = 'import os; os.system("rm -rf /")'
    unsafe_result = bridge.validate_and_execute(unsafe_code, {"language": "python"})
    print(f"Unsafe Code: {unsafe_code}")
    print(f"Result: {json.dumps(unsafe_result, indent=2)}\n")
    
    # Test AutoGen integration (if available)
    try:
        print("--- Testing AutoGen Integration ---")
        config_list = create_local_config()
        integration = CursorAutoGenIntegration(config_list, bridge)
        
        # Test code generation and validation
        task = "Create a function to log messages with timestamps"
        result = integration.generate_and_validate_code(task)
        print(f"Task: {task}")
        print(f"Generated Code: {result['generated_code'][:100]}...")
        print(f"Validation Status: {result['validation_result']['status']}")
        
    except ImportError:
        print("AutoGen not available - skipping integration test")
    except Exception as e:
        print(f"AutoGen integration error: {e}")
    
    # Get statistics
    print("\n--- Getting Statistics ---")
    stats = bridge.get_validation_statistics()
    print(f"Statistics: {json.dumps(stats, indent=2)}")

if __name__ == "__main__":
    main()
