requirements.txt
# ID: [WOLFIE_AGI_UI_REQUIREMENTS_20250923_001]
# SUPERPOSITIONALLY: [dream_data_analysis, quantum_tabs, multi_agent_coordination, bridge_crew_tracking, requirements, dependencies, python_packages]
# DATE: 2025-09-23
# TITLE: requirements.txt — Python Dependencies
# WHO: WOLFIE (Eric) - Project Architect & Dream Architect
# WHAT: Complete list of Python dependencies for WOLFIE AGI UI project
# WHERE: C:\START\WOLFIE_AGI_UI\
# WHEN: 2025-09-23, 11:45 AM CDT (Sioux Falls Timezone)
# WHY: Ensure all required Python packages are available for the project
# HOW: pip install -r requirements.txt
# HELP: Contact WOLFIE for dependency issues or package conflicts
# AGAPE: Love, patience, kindness, humility in dependency management

# Core Flask and Web Framework
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0

# Database and Data Processing
pandas==2.1.1
sqlite3==0.4.6

# Cryptography and Security
cryptography==41.0.4
Fernet==1.0.1

# Data Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
networkx==3.1

# HTTP Requests and API
requests==2.31.0
urllib3==2.0.4

# Configuration and Utilities
configparser==5.3.0
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# JSON and Data Serialization
jsonschema==4.19.0

# Logging and Monitoring
loguru==0.7.0

# Email (for convergence protocol)
secure-smtplib==0.1.1

# Threading and Concurrency
threading==1.0.0

# File and Path Operations
pathlib==1.0.1

# Regular Expressions
regex==2023.8.8

# Math and Statistics
numpy==1.24.3
scipy==1.11.1

# WebSocket Support
websocket-client==1.6.3

# Development and Testing
pytest==7.4.0
pytest-flask==1.2.0

# Code Quality
flake8==6.0.0
black==23.7.0

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Optional: Machine Learning (for future enhancements)
# scikit-learn==1.3.0
# tensorflow==2.13.0
# torch==2.0.1

# Optional: Advanced Visualization
# bokeh==3.2.2
# dash==2.11.1

# Optional: Database ORM (for future enhancements)
# SQLAlchemy==2.0.19
# alembic==1.11.1

# Optional: Caching (for performance)
# redis==4.6.0
# memcached==1.6.3

# Optional: Task Queue (for background tasks)
# celery==5.3.1
# rq==1.15.1

# Optional: Monitoring and Metrics
# prometheus-client==0.17.1
# grafana-api==1.0.3

# Optional: Cloud Integration
# boto3==1.28.25
# azure-storage-blob==12.17.0
# google-cloud-storage==2.10.0
